# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "The transitIOM REST API",
        "title": "Project Mercury API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Jayden Thompson",
            "email": "admin@transitIOM.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.1.0"
    },
    "host": "localhost:8090",
    "basePath": "/api/v1",
    "paths": {
        "/messages/": {
            "get": {
                "description": "Retrieves the last 3 lines of the message log and the version ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "messages"
                ],
                "summary": "Get latest messages",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.GetMessagesResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Uploads a new message to the message log.",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "messages"
                ],
                "summary": "Upload a new message",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Message content",
                        "name": "message",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/api.PutMessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            }
        },
        "/messages/version": {
            "get": {
                "description": "Retrieves the version ID of the latest message log.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "messages"
                ],
                "summary": "Get latest message log version ID",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.GetVersionIDResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            }
        },
        "/schedule/": {
            "get": {
                "description": "Retrieves the download URL and version ID of the latest GTFS schedule.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "schedule"
                ],
                "summary": "Get latest GTFS schedule download URL",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.GetTimetableResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Uploads a new GTFS schedule (zip file).",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "schedule"
                ],
                "summary": "Upload a new GTFS schedule",
                "parameters": [
                    {
                        "type": "file",
                        "description": "GTFS Schedule zip file",
                        "name": "GTFSSchedule",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/api.PutTimetableResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            }
        },
        "/schedule/version": {
            "get": {
                "description": "Retrieves the version ID of the latest GTFS schedule.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "schedule"
                ],
                "summary": "Get latest GTFS schedule version ID",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.GetVersionIDResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 500
                },
                "message": {
                    "type": "string",
                    "example": "Internal Server Error"
                }
            }
        },
        "api.GetMessagesResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "messages": {
                    "type": "string",
                    "example": "{\"timestamp\": \"2026-1-1T00:00:00.000Z\", \"message\": \"Example message\"}"
                },
                "versionID": {
                    "type": "string",
                    "example": "5e4b7d12-542f-4ecf-8d95-7fbec7f7e806"
                }
            }
        },
        "api.GetTimetableResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "downloadURL": {
                    "type": "string",
                    "example": "https://example.com/GTFSSchedule.zip"
                },
                "versionID": {
                    "type": "string",
                    "example": "20231215-143022"
                }
            }
        },
        "api.GetVersionIDResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "version": {
                    "type": "string",
                    "example": "20231215-143022"
                }
            }
        },
        "api.PutMessageResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 202
                },
                "versionID": {
                    "type": "string",
                    "example": "5e4b7d12-542f-4ecf-8d95-7fbec7f7e806"
                }
            }
        },
        "api.PutTimetableResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 202
                },
                "versionID": {
                    "type": "string",
                    "example": "20231215-143022"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}